move taggedarray back to solvers
