+++ src/moscot/problems/__init__.py
@@ -1,4 +1,3 @@
-from moscot.problems._tagged_array import Tag, TaggedArray
 import moscot.datasets
 import moscot.problems.time
 import moscot.problems.space
